backend/pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = "API for gestión de tareas"
authors = ["Your Name <nicolashurtado0712@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "4.2.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
django-cors-headers = "3.14.0"
psycopg2-binary = "2.9.6"
gunicorn = "20.1.0"
python-dotenv = "1.0.0"
dj-database-url = "2.0.0"
django-filter = "23.2"
drf-spectacular = "0.26.5"
pyjwt = "2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.3.1"
pytest-django = "4.5.2"
pytest-cov = "4.1.0"
coverage = "7.2.5"
black = "23.3.0"
flake8 = "6.0.0"
flake8-pyproject = "1.2.3"
mypy = "1.3.0"
model-bakery = "1.20.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "task_manager.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
django_find_project = true
testpaths = ["."]
addopts = "--cov=. --cov-report=term-missing --no-cov-on-fail"

[tool.coverage.run]
source = ["."]
omit = [
    "*/migrations/*",
    "*/tests.py",
    "*/admin.py",
    "*/apps.py",
    "*/urls.py",
    "*/wsgi.py",
    "*/asgi.py",
    "*/conftest.py",
    "*/venv/*",
    "*/management/commands/*",
    "manage.py",
    "wait_for_db.py",
    "create_superuser.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError"
]

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 120
exclude = [
    ".git",
    "__pycache__",
    "docs/source/conf.py",
    "old",
    "build",
    "dist",
    "venv",
    ".venv",
    "*/admin.py",
    "create_superuser.py"
]
extend-ignore = ["E203"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
exclude = ["venv/"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "django.*"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "task_manager.settings" 